# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SamudraKarsa Work Space API
  version: 1.1.0-draft
  summary: Kontrak REST — sebagian ditulis sebelum implementasi, sebagian sesudahnya.
  description: |
    Kontrak REST untuk backend `SamudraKarsaWorkSpace-api`.

    ## Soal urutan penulisan — dan satu deviasi yang dicatat terus terang

    `PRD-BACKEND.md` §12 Fase 1 meminta kontrak ini selesai **sebelum** satu
    baris implementasi. Untuk `health`, `auth`, `work-items`, dan `requests`,
    urutan itu dipatuhi. Untuk `organization`, `profiles`, dan
    `POST /requests/{id}/sync-conflict`, **tidak** — ketiganya baru ditulis
    setelah modulnya dibangun di Fase 3.

    Itu deviasi yang diketahui, dan ia dicatat di sini alih-alih ditutupi.

    Yang hilang karenanya nyata: manfaat "memaksa keputusan diambil saat masih
    murah". Bagian-bagian itu adalah **deskripsi atas apa yang sudah ada**,
    bukan perancangan atas apa yang akan ada — menamai field dan memilih status
    code-nya sudah tidak gratis lagi, karena kodenya sudah bergantung padanya.

    Yang **tidak** hilang, dan inilah alasan bagian-bagian itu tetap ditulis:
    berkas ini satu-satunya tempat bentuk API bisa ditinjau sebagai satu
    kesatuan. Tanpa itu, `organization` dan `profiles` hanya bisa diketahui
    dengan membaca controller-nya satu per satu.

    ## Kenapa berkas ini ditulis tangan

    `PRD-BACKEND.md` §11 menetapkan bahwa dokumentasi yang **disajikan** di
...

...
    bukan sumbernya lagi.

...

    ## Cakupan berkas ini

...

...

...

...
servers:
...
security:
  - bearerAuth: []

# ═══════════════════════════════════════════════════════════════════════════
# Jalur
# ═══════════════════════════════════════════════════════════════════════════
paths:
...
  /auth/google:
    post:
      tags: [auth]
      operationId: loginWithGoogle
      summary: Masuk dengan Google
      description: |
        Jalur utama. BFF menukar kode OAuth dengan **ID token** di sisinya, lalu
        meneruskan token itu apa adanya ke sini.

        **Yang diterima adalah ID token mentah, bukan klaim yang sudah diurai.**
        Versi awal `PRD-REVAMP.md` §7.8.3 menggambarkan BFF mengirim `email`,
        `email_verified`, dan `sub` sebagai JSON biasa — dan bentuk itu tidak
        bisa diverifikasi apa pun, sehingga
        `{"email": "<email owner>", "emailVerified": true, "sub": "apa saja"}`
        menghasilkan sesi owner tanpa satu kredensial pun. ID token adalah JWT
        yang ditandatangani Google, jadi klaimnya bisa diperiksa. Lihat
        `GoogleLoginRequest`.

        Backend **tidak pernah bicara ke Google**: tidak ada penukaran kode, dan
        tidak ada `GOOGLE_CLIENT_SECRET` di sisi ini.

        **Lima pemeriksaan tertutup** (`PRD-REVAMP.md` §7.8.2), semuanya wajib:

        1. `email_verified` harus `true` — dibaca dari token yang **sudah
           terverifikasi**, bukan dari isian pemanggil
        2. Email dinormalisasi di kedua sisi
        3. **Tidak ada pembuatan akun otomatis** — profil harus sudah ada
        4. Status selain `inactive` diterima; `inactive` **ditolak**
        5. `sub` disimpan; pencocokan berikutnya lewat `sub`, bukan `email`

        Nomor 3 adalah inti "tertutup". Nomor 5 menutup skenario langka tetapi
        nyata: email lama dipakai ulang orang lain, dan orang itu masuk ke akun
        anggota sebelumnya. `sub` yang tersimpan berbeda dari `sub` yang dikirim
        **ditolak**, bukan diperbarui.

        `invited` menjadi `active` pada login pertama yang berhasil.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GoogleLoginRequest' }
      responses:
        '200':
          description: Berhasil masuk.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '401':
          description: |
            Email tidak terdaftar, akunnya nonaktif, `sub`-nya tidak cocok, atau
            ID tokennya tidak sah.

            Pesannya **selalu sama** untuk semua sebab itu, dan sama juga dengan
            kegagalan di `/auth/login` (`PRD-REVAMP.md` §7.8.4). Pesan yang
            berbeda memberi tahu penyerang email mana yang terdaftar — dan itu
            membatalkan sifat tertutup sistem ini.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/login:
    post:
      tags: [auth]
      operationId: loginWithPassword
      summary: Masuk dengan password
      description: |
        Jalur cadangan, untuk anggota yang tidak bisa atau tidak mau memakai
        Google. Password diperiksa dengan **argon2id**, bukan bcrypt.

        Memanggil `canAuthenticate(email)` yang **sama persis** dengan yang
        dipanggil jalur Google (`PRD-BACKEND.md` §6.1). Dua jalur masuk berarti
        dua tempat aturan "tertutup" ditegakkan; satu fungsi berarti tidak ada
        yang bisa menyimpang.

        Kalau passwordnya adalah password sementara dari owner,
        `mustChangePassword` menyala di hasilnya — dan sejak saat itu backend
        menolak seluruh endpoint kecuali jalur ganti-password
        (`PRD-BACKEND.md` §6.5).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordLoginRequest' }
      responses:
        '200':
          description: Berhasil masuk.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '401':
          description: |
            Kredensial tidak sah, email tidak terdaftar, atau akun nonaktif —
            satu pesan untuk ketiganya.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refreshSession
      summary: Memperbarui access token
      description: |
        Refresh token **dirotasi setiap dipakai**: yang lama langsung dicabut dan
        digantikan yang baru (`PRD-REVAMP.md` §7.8.7).

        Rotasi bukan kerapian. Refresh token yang tidak dirotasi berlaku sampai
        kedaluwarsa tanpa cara mengetahui bahwa ia sudah disalin orang lain —
        sedangkan dengan rotasi, pemakaian token lama adalah bukti bahwa ada dua
        pemegang, dan keluarga token itu bisa dicabut seluruhnya.

        Refresh token **tidak pernah** dikirim di badan permintaan: ia ada di
        cookie `httpOnly` bernama `sk_refresh` yang dipegang BFF (keputusan 31).
        BFF meneruskan cookie itu apa adanya sebagai header `Cookie`.

        Jalur lewat badan permintaan sengaja **tidak** disediakan. Ia akan
        menjadi rute kedua bagi token paling sensitif di sistem ini — dan rute
        kedua itu adalah rute yang isinya ikut tercatat di log permintaan.
      security: []
      responses:
        '200':
          description: Token baru.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '401':
          $ref: '#/components/responses/NotAuthenticated'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/session:
    get:
      tags: [auth]
      operationId: getSession
      summary: Siapa yang sedang masuk
      description: |
        Mengembalikan aktor yang sedang masuk, dari tokennya.

        **Tidak publik**, dan tetap boleh dipakai saat `mustChangePassword`
        menyala — ia ada di `PENDING_PASSWORD_EXEMPT`. Tanpa pengecualian itu,
        frontend tidak bisa mengetahui bahwa yang perlu ditampilkan adalah layar
        ganti password: ia hanya akan menerima penolakan tanpa tahu sebabnya.
      responses:
        '200':
          description: Aktor yang sedang masuk.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Actor' }
        '401':
          $ref: '#/components/responses/NotAuthenticated'
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Keluar
      description: |
        Mencabut refresh token yang sedang dipakai. Access token yang masih
        berlaku **tidak** bisa dicabut — ia mati sendiri dalam 15 menit.
        Itu batas yang diterima, bukan yang terlupa.
      responses:
        '204':
          description: Sesi dicabut.
        '401':
          $ref: '#/components/responses/NotAuthenticated'
  /auth/change-password:
    post:
      tags: [auth]
      operationId: changePassword
      summary: Mengganti password sendiri
      description: |
        Satu-satunya jalan keluar saat `mustChangePassword` menyala.

        Dua akibat yang wajib menyertainya (`PRD-REVAMP.md` §7.8.6):

        - Penanda `mustChangePassword` dimatikan
        - **Seluruh refresh token anggota itu dicabut.** Sesi lama tidak boleh
          bertahan setelah password berubah — kalau bertahan, mengganti password
          bukan tindakan pengamanan, melainkan tindakan yang tidak mengubah apa
          pun bagi orang yang sudah memegang sesinya.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '204':
          description: Password diganti, sesi lain dicabut.
        '401':
          $ref: '#/components/responses/NotAuthenticated'
        '409':
          description: Password saat ini tidak cocok.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          $ref: '#/components/responses/ValidationFailed'

  # ─────────────────────────────────────────────────────────────────────────
  # Pekerjaan
  # ─────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token JWT, umur 15 menit, berisi `sub`, `roles`, `divisionCodes`,
        dan `jti`.

        `roles` berupa array (keputusan 52), bukan satu nilai — sehingga token
        tidak perlu diterbitkan ulang saat seseorang ditambahkan sebagai kepala
        divisi di divisi lain. Klaimnya adalah **salinan** keadaan saat token
        dibuat; sumber kebenarannya tetap `profiles`, dan umur 15 menit adalah
        batas berapa lama salinan itu boleh basi.

        Dikirim **oleh BFF**, bukan oleh browser (keputusan 31). Token tidak
        pernah menyentuh JavaScript: BFF menyimpannya di cookie `httpOnly` dan
        menukarnya menjadi header ini saat memanggil backend. Konsekuensinya
        untuk kontrak ini: tidak ada satu pun operasi di sini yang boleh
        mengandalkan cookie, dan CORS boleh menolak seluruh origin browser
        (`PRD-SISTEM.md` §5).
...
  responses:
    NotAuthenticated:
      description: Token tidak ada, kedaluwarsa, atau tidak sah.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.invalid/problems/not-authenticated
            title: Permintaan ini memerlukan token yang sah
            status: 401
            code: not_authenticated
            instance: /api/v1/work-items
            requestId: 01J8XQ7K3M9W2P
...
    ValidationFailed:
      description: |
        Isian tidak lolos aturan (`PRD-REVAMP.md` §7.5).

        `422`, bukan `400`. Perbedaannya berguna: `400` berarti ada yang salah
        di kode pemanggil, `422` berarti ada yang salah di isian pengguna — dan
        hanya yang kedua yang memang untuk ditampilkan.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
          example:
            type: https://api.invalid/problems/validation-failed
            title: Data yang dikirim tidak valid
            status: 422
            code: validation_failed
            instance: /api/v1/work-items
            requestId: 01J8XQ7K3M9W2P
            errors:
              - field: dueDate
                code: invalid_format
                message: Tanggal tidak dikenali.
...
    RateLimited:
      description: |
        Melewati batas percobaan. Login 5 per 15 menit; tulis 100/menit; baca
        300/menit; ekspor 5/jam.

        Penghitungnya ada di tabel `rate_limit_counters` di Postgres, bukan di
        memori proses (keputusan 43). Penghitung in-memory akan tampak berfungsi
        saat diuji dan gagal saat dipakai: setiap instans serverless punya
        memorinya sendiri, sehingga batas sebenarnya adalah batas dikali jumlah
        instans.
      headers:
        Retry-After:
          description: Detik sampai boleh mencoba lagi.
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      description: |
        RFC 7807 (`PRD-REVAMP.md` §7.5). Seluruh error berbentuk ini — tidak ada
        controller yang menyusun bentuk errornya sendiri, karena bentuk yang
        bersaing berarti frontend harus menebak.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: |
            URI yang menunjuk ke penjelasan jenis masalah ini. Dibentuk dari
            `API_PUBLIC_URL` (keputusan 38), sehingga pindah domain tidak
            menuntut apa pun ditulis ulang.
        title:
          type: string
          description: Ringkasan singkat dan stabil. Boleh ditampilkan ke pengguna.
        status: { type: integer, minimum: 400, maximum: 599 }
        detail:
          type: string
          description: |
            Penjelasan kejadian ini. **Tidak pernah memuat detail internal** —
            nama tabel, potongan SQL, jejak tumpukan. Pada `500`, field ini
            dihilangkan sama sekali.
        instance: { type: string, description: Path request yang gagal. }
        requestId:
          type: string
          description: |
            Untuk menghubungkan keluhan dengan log.

            Berguna **hanya dalam satu jam**: retensi log Vercel adalah 1 jam
            (`PRD-REVAMP.md` §6.2). Di luar rentang itu ia tidak menunjuk apa
            pun, dan itu batas yang harus disadari, bukan ditemukan saat
            dibutuhkan.
        code:
          type: string
          description: |
            Kode mesin yang stabil, untuk dipetakan frontend. Berbeda dari
            `title` yang boleh berubah kata-katanya, `code` tidak pernah berubah
            artinya.
      additionalProperties: true
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          required: [errors]
          properties:
            code: { type: string, const: validation_failed }
            errors:
              type: array
              description: Satu entri per field yang gagal.
              items: { $ref: '#/components/schemas/ValidationError' }
    ValidationError:
      type: object
      required: [field, code, message]
      properties:
        field:
          type: string
          description: |
            Jalur field, bertitik untuk yang bersarang (`extraFields.accountNumber`).
            Kosong untuk kegagalan yang tidak terikat field tertentu.
        code:
          type: string
          description: Kode mesin dari pemeriksaannya, mis. `invalid_format`, `too_small`.
        message:
          type: string
          description: Kalimat bahasa Indonesia untuk ditampilkan di sebelah isiannya.
...
    Role:
      type: string
      description: |
        Lima peran (keputusan 46), bukan empat. `division_deputy` memegang
        **seluruh** hak `division_head` untuk sekarang; ia berdiri sendiri supaya
        wewenangnya bisa dipisah nanti tanpa migrasi data.

        Label yang dilihat pengguna tetap "Kadiv" dan "Wakil Kadiv" — dictionary
        frontend yang menerjemahkannya, bukan backend.
      enum: [owner, co_owner, division_head, division_deputy, member]
...
    Actor:
      type: object
      description: Siapa yang sedang masuk.
      required: [id, email, fullName, roles, mustChangePassword]
      properties:
        id:
          type: string
          format: uuid
          description: '`profiles.id` — bukan id di penyedia identitas.'
        email: { type: string, format: email }
        fullName:
          type: [string, 'null']
          description: |
            Nama lengkap, untuk ditampilkan. Selalu ada di respons —
            `profiles.full_name` `not null` — tetapi boleh `null` di kontrak
            karena baris lama dari `sksks` bisa saja kosong.
        roles:
          type: array
          items: { $ref: '#/components/schemas/Role' }
          description: |
            Array, bukan satu nilai (keputusan 52) — satu orang bisa menjadi
            kepala divisi sekaligus anggota di divisi lain.
        divisionCodes:
          type: array
          items: { type: string }
          description: |
            Kode divisi yang **dinaungi** — terisi hanya untuk `division_head`
            dan `division_deputy`. Anggota biasa di divisi Sekretaris &
            Bendahara **tidak** mengisi ini, dan justru itu yang membuat
            pemeriksaan keuangan benar.
        mustChangePassword:
          type: boolean
          description: |
            Kalau true, **seluruh endpoint selain jalur ganti-password ditolak**.
            Penegakannya di backend, bukan di UI: kalau hanya frontend yang
            menahan, satu `curl` dengan token yang sah sudah cukup melewatinya
            (keputusan 10).
        mustChangePasswordExempt:
          type: array
          items: { type: string }
          description: Policy yang tetap boleh dipakai saat penanda itu menyala.

    # ───────────────────────────────────────────────────────────────────────
    # Health
    # ───────────────────────────────────────────────────────────────────────
...
    GoogleLoginRequest:
      type: object
      required: [credential]
      properties:
        credential:
          type: string
          writeOnly: true
          description: |
            **ID token Google, apa adanya.** Ini yang diterima callback OAuth
            sebagai `id_token`, dan BFF meneruskannya tanpa mengurai apa pun.

            Backend memverifikasi tanda tangannya terhadap kunci publik Google,
            `aud`-nya terhadap `GOOGLE_CLIENT_ID`, `iss`-nya terhadap
            `accounts.google.com`, dan `exp`-nya. Klaim di dalamnya baru dibaca
            **setelah** keempatnya lolos.

            Yang dibaca dari token yang sudah terverifikasi itu:

            - `email_verified` — wajib `true`. Kalau penyedia pernah
              mengembalikan email yang belum terverifikasi, orang bisa
              mengklaim email orang lain, dan itu membatalkan sifat tertutup
              sistem ini.
            - `sub` — pengenal tetap, bukan email. Email bisa berubah; `sub`
              stabil. Setelah login pertama, pencocokan lewat sini.
            - `email` — dinormalisasi (huruf kecil, tanpa spasi) di kedua sisi.

            Bentuk lama `{email, emailVerified, sub}` sebagai JSON biasa **tidak
            didukung**, dan tidak boleh dikembalikan: ketiganya teks yang bisa
            ditulis siapa saja, sehingga siapa pun bisa membuat sesi owner tanpa
            satu kredensial pun.
    PasswordLoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1, writeOnly: true }
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, writeOnly: true }
        newPassword:
          type: string
          minLength: 12
          writeOnly: true
          description: Minimal 12 karakter. Diperiksa dengan argon2id.
    AuthResult:
      type: object
      required: [actor, accessToken, refreshToken, expiresIn]
      properties:
        actor: { $ref: '#/components/schemas/Actor' }
        accessToken:
          type: string
          writeOnly: true
          description: |
            JWT, umur 15 menit, berisi **hanya `sub`**.

            Peran, status keanggotaan, dan `mustChangePassword` sengaja **tidak**
            ada di dalamnya: klaim adalah salinan keadaan saat token dibuat, dan
            salinan itu berlaku sampai tokennya kedaluwarsa. Anggota yang
            dinonaktifkan akan tetap bisa memakai sistem selama sisa umur
            tokennya, peran yang diturunkan tetap berlaku, dan
            `mustChangePassword` yang menyala tidak terlihat. Backend memuat
            aktornya dari database pada setiap permintaan, sehingga ketiganya
            selalu keadaan sekarang.

            Tanpa `jti`: ia hanya berguna kalau ada daftar cabut access token,
            dan yang ada adalah umur 15 menit.
        refreshToken:
          type: string
          writeOnly: true
          description: |
            **Diteruskan BFF ke cookie `httpOnly`; tidak pernah menyentuh
            JavaScript.** Umur 7 hari, dan **dirotasi setiap kali dipakai**.

            Backend mengembalikannya di badan respons — bukan lewat `Set-Cookie`
            — karena yang memegang cookie adalah BFF (keputusan 31), dan
            backend tidak pernah berbicara langsung dengan browser.

            Yang disimpan di database adalah **SHA-256 dari token ini**, bukan
            tokennya. Isi tabel yang bocor tidak langsung memberi penyerang sesi
            siapa pun.
        expiresIn: { type: integer, description: 'Detik. 900 untuk access token.' }

    # ───────────────────────────────────────────────────────────────────────
    # Kolom bersama
    # ───────────────────────────────────────────────────────────────────────
...